requirements.txt
torch==1.12
pyyaml
easydict
numpy
tqdm
scipy
tensorboard
matplotlib
scikit-learn
h5py
open3d==0.15
einops